Move remaining SFBAudioEngine seams behind the engine abstraction
Add MediaBackend, a pure selector that reports the chosen engine while each
facade picks its own backend, and decouple metadata extraction into a
MetadataEngine facade with the SFB parsing isolated in SFBMetadataReader.
Group the seams under Core/Playback and Core/Metadata, consolidate Now Playing
into a single PlaybackManager-owned path, and drop the vestigial SFB import in
DMTrackUpdate. Runtime stays SFB only; this just localizes engine knowledge so
adding Crescendo (and later removing SFB) is a small, contained change.
